# perf: boost ECDH key creation for Primitive execution